Adding detail, fixing some bugs

+++ includes/widget/jgl_hour_entry.h
@@ -28,7 +28,7 @@ public:
 
 	void round();
 
-	int value(){return (_hour.value() * 60 + _minute.value());}
+	float value(){return (_hour.value() * 60 + _minute.value());}
 
 	bool selected(){return (_selected);}
 	w_box_component &box(){return (_box);}
